requirements.txt
# Core data handling and numerical libraries
pandas>=1.5.0
numpy>=1.21.0

# Statsmodels for econometric modeling (OLS, VAR, VECM, SARIMAX)
statsmodels>=0.13.0

# Scipy for scientific computing utilities
scipy>=1.8.0

# Scikit-learn for ML preprocessing, baseline models, metrics, CV
scikit-learn>=1.1.0

# Visualization libraries
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.9.0
kaleido>=0.2.0  # For static plotly exports

# Configuration and environment variables
pyyaml>=6.0
python-dotenv>=0.20.0

# Data fetching
requests>=2.28.0

# Time Series Specific Libraries
pmdarima>=2.0.0
arch>=5.0.0

# ML Libraries
xgboost>=1.7.0
lightgbm>=3.3.0
shap>=0.41.0

# Notebook environment
jupyterlab>=3.6.0
ipykernel>=6.0.0

# Code quality
black>=22.0.0
flake8>=6.0.0
pytest>=7.0.0
pmdarima>=2.0.4
